backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.1.0
reportlab==4.0.7
gunicorn==21.2.0
whitenoise==6.6.0
geopy==2.4.0
polyline==2.0.0
drf-spectacular==0.27.0